Revert "202 add more cloudcdnwaf provider to ip detection (#205)"
This reverts commit 32f54cd051e64914d85f8a9275bc764be30fdbdc.
